Pushing updated SPD Attribute values
Per Kevin and Dan, the DDIMM spec indicates a 2-byte address, which means this needs to change.
            <field><id>byteAddrOffset</id><value>0x01</value></field>		--> 0x2

It is a 32Kbit (=4KB) part so this is wrong:
            <field><id>maxMemorySizeKB</id><value>0x01</value></field>		---> 0x4

PAGE WRITE: The 32K/64K EEPROM is capable of 32-byte page writes.
            <field><id>writePageSize</id><value>0x50</value></field>		--> 0x20

Self-Timed Write Cycle (10 ms max)
            <field><id>writeCycleTime</id><value>0x05</value></field>		--> looks good, we have a NACK retry that will cover the worst case

What voltage are we running at?  I think it is 1.8V?  So we need to have the bus frequency set to 100KHz.