Adjust PBAX_GROUP id and PROC_ERROR_TEMP settings to get OCC running
The OCC does not support having CHIP_AS_GROUP numbering for PBAX
GROUP/CHIP ids. If you look at Witherspoon they have CHIP_AS_GROUP
set for fabric topology yet they are saying PBAX_GROUP is 0 for both
chips and PBAX_CHIP is 1 for proc1. As we know this works in wither-
spoon system we will follow this in Axone. Also adjust the temp that
we callout the processor to be 95 degrees C instead of 5.