Update PCIE bus settings
There was information missing that describes the layout of the
PEC targets in the system. The correct config is to have PEC0
as one X16 set of lanes, PEC1 to be two X8 sets of lanes, and
PEC2 also as one X16 set of lanes. The is the same for both procs
in the system.