Update DL_POS and REFCLOCK_SWIZZLE for OMI targets
When trying to bring up a system that had an OCMB behind the 2nd
processor for the first time we found that some of the attributes
that describe the relationship of the OMI to the OMICs and OCMBs
were incorrect. This commit addresses those issues and allows fw
to correctly identify the connection we need to enable in order
to activate all of the ocmbs, not just the ocmb behind the first
proc in the first slot.