Update swift MRW with values found to be needed during explorer bup
Over the course of the explorer bringup we found that we needed some
updates to the swift xml. The first issue we hit is the the i2c addresses
for the pmics on the explorers were incorrect. In the future the fw
will read these values from the SPD but for now we need them set in
the mrw correctly. Additionally the mrw incorrectly had pmic1's
position set to 0 when it needed to be 1. The other changes included
setting the PRBS delay time which is used during the OMI link trainig
workarounds and setting IO_OMI_TX_FFE_PRECURSOR to be 0x0B to match
what cronus has set. I also noticed that we were missing the NOT_WIRED
enumeration for the OPTICS_CONFIG_MODE value so I added that in also.